LayoutTests/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-import-csp-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'worker-src'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'worker-src'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'worker-src'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'worker-src'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'worker-src'.


PASS worker-src 'self' directive should disallow cross origin static import. 
FAIL worker-src * directive should allow cross origin static import. assert_array_equals: property 0, expected "export-on-load-script.js" but got "ERROR"
PASS script-src 'self' directive should disallow cross origin static import. 
FAIL script-src * directive should allow cross origin static import. assert_array_equals: property 0, expected "export-on-load-script.js" but got "ERROR"
FAIL worker-src * directive should override script-src 'self' directive and allow cross origin static import. assert_array_equals: property 0, expected "export-on-load-script.js" but got "ERROR"
PASS worker-src 'self' directive should override script-src * directive and disallow cross origin static import. 
PASS script-src 'self' directive should disallow cross origin dynamic import. 
FAIL script-src * directive should allow cross origin dynamic import. assert_array_equals: property 0, expected "export-on-load-script.js" but got "ERROR"
FAIL worker-src 'self' directive should not take effect on dynamic import. assert_array_equals: property 0, expected "export-on-load-script.js" but got "ERROR"
